BuildingSystem rotation implementation and new City function
We still have a bazillion bugs to work on.

For now, we have successful proof that this does, in fact, work vaguely on Unity.

Moved all of the files to an old file and created new C# files that will dictate the overall underlying logic and systems of the entire game. The GUI will sit on top of these